Update code/codium marketplace change scripts
Now checks multiple file paths, and alerts which files were changed. On
failure it returns an error and warns that no files were modified.

+++ linux-install/change-code-marketplace.py
@@ -1,16 +1,30 @@
 #!/usr/bin/env python
 
 import json
+import os
 
-# This works at least for Arch Linux's 'Code - OSS'
+file_paths = [
+    "/usr/lib/code/product.json",
+    "/usr/share/codium/resources/app/product.json"
+]
 
-file_path = "/usr/lib/code/product.json"
+fileChanged = False
 
-with open(file_path, "r") as f:
-    config = json.load(f)
-    config["extensionsGallery"]["serviceUrl"] = "https://marketplace.visualstudio.com/_apis/public/gallery"
-    config["extensionsGallery"]["cacheUrl"] = "https://vscode.blob.core.windows.net/gallery/index"
-    config["extensionsGallery"]["itemUrl"] = "https://marketplace.visualstudio.com/items"
+for path in file_paths:
+    print('Checking for file: {}'.format(path))
+    if os.path.isfile(path):
+        with open(path, "r") as f:
+            config = json.load(f)
+            config["extensionsGallery"]["serviceUrl"] = "https://marketplace.visualstudio.com/_apis/public/gallery"
+            config["extensionsGallery"]["cacheUrl"] = "https://vscode.blob.core.windows.net/gallery/index"
+            config["extensionsGallery"]["itemUrl"] = "https://marketplace.visualstudio.com/items"
 
-with open(file_path, "w") as f:
-    json.dump(config, f, indent=2)
+        with open(path, "w") as f:
+            json.dump(config, f, indent=2)
+
+        print('  Changed marketplace in {}'.format(path))
+        fileChanged = True
+
+if not fileChanged:
+    print('Warning: No marketplaces could be found to be changed')
+    exit(1)
